requirements.txt
Scrapy==2.11.2
scrapy-rotating-proxies==0.6.2
python-dotenv==1.0.1
